Rename RunnerUtil and move Console runner into its own folder
